requirements-dev.txt
# Development dependencies for Lippmann-Rau Archive OCR System
# Install with: pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# Code formatting
black>=23.0.0
isort>=5.12.0

# Linting
flake8>=6.0.0
pylint>=2.17.0
mypy>=1.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development tools
ipython>=8.12.0
jupyter>=1.0.0

# Data visualization (for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Type stubs
pandas-stubs>=2.0.0
types-requests>=2.28.0
